requirements.txt
python-telegram-bot>=20.0
anthropic>=0.39.0
python-dotenv>=1.0.0
httpx>=0.27.0
cryptography>=43.0.0

# Vector embeddings & semantic search
chromadb>=0.5.0

# Image generation
google-genai>=1.0.0

# Document processing
python-docx>=1.0.0

# DeFi Trading
web3>=7.0.0
eth-account>=0.13.0

# Research & trend monitoring
pytrends
arxiv
praw
beautifulsoup4
lxml
